feat(config): add option to control loaded config debug log printing
Add a new configuration option `printLoadedDebugLog` that allows
controlling whether the loaded configuration is printed in the debug
log. The default value is set to true to maintain backward compatibility.

The new option can be configured using the `WithPrintLoadedDebugLog`
function which accepts a boolean parameter to enable or disable the
debug log printing of loaded configurations.